requirements.txt
setuptools~=67.8.0
beets~=1.6.0
python-mpd~=0.3.0
confuse~=2.0.1
